SprintBoot User Track List
This project is a Spring Boot application designed for managing user data. It allows users to create, view, and list user data through a web interface.  The application tracks the number of page hits and displays saved user data using Thymeleaf templates. Data is stored in an H2 in-memory database.